AsyncCompletion: Build fix
svn:sync-xref-src-repo-rev: r75518

+++ src/VBox/VMM/VMMR3/PDMAsyncCompletionFile.cpp
@@ -1,4 +1,4 @@
-/* $Id: PDMAsyncCompletionFile.cpp 39680 2011-12-27 21:18:27Z alexander.eichner@oracle.com $ */
+/* $Id: PDMAsyncCompletionFile.cpp 39681 2011-12-27 21:51:07Z alexander.eichner@oracle.com $ */
 /** @file
  * PDM Async I/O - Transport data asynchronous in R3 using EMT.
  */
@@ -1094,7 +1094,7 @@ static int pdmacFileEpRead(PPDMASYNCCOMPLETIONTASK pTask,
     LogFlowFunc(("pTask=%#p pEndpoint=%#p off=%RTfoff paSegments=%#p cSegments=%zu cbRead=%zu\n",
                  pTask, pEndpoint, off, paSegments, cSegments, cbRead));
 
-    if (RT_UNLIKELY(off + cbRead > pEpFile->cbFile))
+    if (RT_UNLIKELY((uint64_t)off + cbRead > pEpFile->cbFile))
         return VERR_EOF;
 
     STAM_PROFILE_ADV_START(&pEpFile->StatRead, Read);
